refactoring event item to resolve code duplication
